fix Debian initramfs -> disable resume
resume (to swap partition) is now disabled for up coming
update-initramfs commands

probleme was that the UUID of the swap partition change, hence initramfs
on other machines may complain.

removing the file or setting RESUME=auto in the clean step seems not
relevant in the context.

see man initramfs.conf