[from_scratch/debian+ubuntu] add variables: deb_mirror_directory deb_security_hostname deb_security_directory and set in preseed
for Debian this has no effect: the security variables are commented in
the preseed files

for ubuntu/arm64, this may be necessary because the security repo is not
in the same place as for x86, but if apt-setup/security_* is commented
in the preseed, it is guessed correctly.

adding variables anyway, this is harmless.