[big] [puppet] [nvidia] Add nvidia-smi service to create /dev/nvidiaX when required
Both cuda and prometheus have services that need to start only on machines
with GPU.  They rely on /dev/nvidia0 being present on such machines to
detect the presence of GPU.

However, /dev/nvidia0 is not created when the nvidia driver is loaded, but
only when `nvidia-smi` is first run.

The new service calls `nvidia-smi` once during boot to create /dev/nvidia0
and thus to allow the other services to start.  On machines without GPU,
`nvidia-smi` will fail, so we ignore its exit code.