[std] raise oar-node service timeout to 5 minutes and remove sh options
This fix issues on some clusters where where multiple run of g5k-checks
are required, this was not possible because of the '-eu' which was
causing oar-node-service to exit prematurely.
Raising the systemd service timeout is also necessary, as g5k-checks was
hanging waiting for remote file system to become available.