[std] remove oar-node.service timeout override
oar-node.service used to start OAR's ssh server and oar-node-service
script which launch g5k-checks and put the node Alive.
Now, with OAR 2.5.10~g5k29, this is splitted in two services:
    - oar-node.service to start OAR's ssh server
    - oar-node-script.service to run the script in charge of
      g5k-checks and oarnodesettings

oar-node-script.service is a 'oneshot' service, meaning it doesn't have
a default timeout. Since the script handles the retries and timeout, we
don't need to override the systemd unit's one.