Fix setup of system fs in chroot
The setup_system_fs_in_chroot macrostep now use aliases, so that the
commands can be used in the checkpoint actions as well. This requires
kameleon ≥ 2.10.16 (so that aliases get resolved in the checkpoint
actions)

The main command alias is `configure-system-fs-chroot_in`, that takes in
charge enabling, setup and unsetup of the system fs in the chroot.

It uses the `setup-system-fs-chroot_in` and `setup-system-fs-chroot_in`
command aliases for setup and unsetup, which can also be used alone.

In particular, when the docker backend's checkpointing is enabled, the
`setup-system-fs-chroot_in` command alias is executed after every
checkpoint creation (required because the docker container goes down and
up during the checkpoint creation, thus it looses the system fs mounts)
or after applying the checkpoint when retarting from a previous
checkpointed build.

This actually setup the chroot's system fs only If the
$KAMELEON_WORKDIR/MOUNT_SYSTEM_FS_IN_CHROOT file was previously created
(done by the `configure_).